More tweaks
- Adding event that is safe for "notifications / triggering" changes.
- Thinking about how to make register work with auto registering
- More notes on challenges
- Bash e2e test for challenge (making testing a little easier)
