+++ server/db/202010291717-challenge.sql
@@ -8,16 +8,11 @@ CREATE TABLE IF NOT EXISTS challenge (
 
 CREATE INDEX IF NOT EXISTS challenge_created ON challenge (user_uuid, created);
 
-CREATE TABLE IF NOT EXISTS challenge_users (
-  uuid CHARACTER(36) not null primary key,
+CREATE TABLE IF NOT EXISTS challenge_records (
+  uuid CHARACTER(36) not null,
   user_uuid CHARACTER(36) not null,
-  info text,
-  UNIQUE(uuid, user_uuid)
+  ext_uuid CHARACTER(36) not null
 );
 
-CREATE TABLE IF NOT EXISTS challenge_records (
-  uuid CHARACTER(36) not null primary key,
-  user_uuid CHARACTER(36) not null,
-  ext_uuid CHARACTER(36) not null,
-  UNIQUE(uuid, user_uuid, ext_uuid)
-);
+CREATE UNIQUE INDEX IF NOT EXISTS challenge_records_uniq ON challenge_records (uuid, user_uuid, ext_uuid);
+
